% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chest_speedglm.R
\name{chest_speedglm}
\alias{chest_speedglm}
\title{Assessing confounding effects using Generalized Linear regression models}
\usage{
chest_speedglm(
  crude,
  xlist,
  data,
  var_lab = "Variables",
  est_lab = "OR (95\% CI)",
  change_lab = "Change, \%",
  family = binomial(),
  digits = "\%.2f",
  digits_change = "\%.1f",
  method = c("eigen", "Cholesky", "qr"),
  maxit = 25,
  epsilon = 1e-08,
  na_omit = FALSE,
  hrzl_lines = gpar(col = "#444444"),
  text_size = 1,
  xlab_size = 1,
  xtick_size = 1,
  xlab = NULL,
  plus = "  + ",
  indicate = FALSE,
  ...
)
}
\arguments{
\item{crude}{An object of \emph{formula} for initial model, generally crude model.
However, any other variables can also be included here as the initial model.}

\item{xlist}{A \emph{vector} of characters with all variable names of potential confounders.}

\item{data}{\emph{Data frame}.}

\item{var_lab}{\emph{Character} string for the column name of \emph{variables} in the graph.}

\item{est_lab}{\emph{Character} string for the column name of effect estimates in the graph.}

\item{change_lab}{\emph{Character} string for the column name of \emph{"Change"} in the graph}

\item{family}{Description of the error distribution. Default is \emph{"binomial"}.}

\item{digits}{Set the display format for number in the graph other than the \emph{"Change"} column. Default: "\%.2f"}

\item{digits_change}{Set the format for the \emph{"Change"} column. Default: "\%.1f"}

\item{method}{Method to detect for singularity.}

\item{maxit}{See \code{'glm'}.}

\item{epsilon}{See \code{'glm'}.}

\item{na_omit}{Remove all missing values.}

\item{hrzl_lines}{A \code{logic} to include or remove horizontal line.}

\item{text_size}{Set text size.}

\item{xlab_size}{Set \code{x} label size.}

\item{xtick_size}{Set \code{x} tick label size.}

\item{xlab}{Add \code{x} label.}

\item{plus}{Change the \code{+} sign before variable names.}

\item{indicate}{indicate progress}

\item{...}{Further optional arguments for forestplot.}
}
\value{
A table with effect estimates and their changes at all steps.
}
\description{
This is a faster option to 'chest_glm'.
It presents the effect estimates (such as \emph{odds ratios}) for the association between exposure and outcome variables by adding other variables (potential confounders) to the model sequentially.
The order of variables to be added is based on the magnitudes of the changes in effect estimates. It returns a graph and a table with effect estimates, 95\% confidence intervals and changes (\%) at different steps.
}
\examples{

vlist<-c("Age", "Sex", "Married", "Cancer", "CVD", "Education", "Income")

chest_speedglm(crude = "Endpoint ~ Diabetes",
              xlist = vlist, na_omit=TRUE, data = diab_df)

 # adding derived terms such as an interaction between Age and Sex, and age squared:

library(tidyverse)
diab_df <- diab_df \%>\%
mutate(Age_Sex = Age*Sex, Age2 = Age^2)

vlist<-c("Age", "Sex", "Age2", "Age_Sex", "Married", "Cancer", "CVD", "Education", "Income")

chest_speedglm(crude = "Endpoint ~ Diabetes", xlist = vlist, na_omit=TRUE, data = diab_df)

}
\seealso{
\pkg{'forestplot'}

\pkg{'speedglm'}

\code{'glm'} of \pkg{'stats'}
}
